+++ package.json
@@ -33,6 +33,7 @@
   },
   "dependencies": {
     "express": "^4.16.4",
-    "if-env": "^1.0.4"
+    "if-env": "^1.0.4",
+    "passport": "^0.4.0"
   }
 }
